requirements.txt
pandas
yfinance
google-adk
pandas-ta
TA-Lib
pandas-datareader
slack_bolt
fastapi
uvicorn[standard]
notion-client
dotenv
aiohttp
requests
pydantic